+++ rapyer/fields/__init__.py
@@ -1,4 +1,5 @@
 from rapyer.fields.index import Index
 from rapyer.fields.key import Key
+from rapyer.fields.safe_load import SafeLoad
 
-__all__ = ["Key", "Index"]
+__all__ = ["Key", "Index", "SafeLoad"]